# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Effettua il login dell'utente",
                "description": "Consente all'utente di effettuare il login con le credenziali (email e password). Se il login è corretto, viene generato un token per l'accesso alle API.",
                "operationId": "userLogin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "remember": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login effettuato con successo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Login effettuato"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "Api token"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Parametri non validi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Parametri non validi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Crea un nuovo utente",
                "description": "Permette la creazione di un nuovo utente. Viene assegnata una password temporanea e inviata un'email all'utente per il settaggio della password.",
                "operationId": "userRegister",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "cf": {
                                        "type": "string",
                                        "example": "ABCDEFGHIJKLMNOP"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "surname": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2001-08-15"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "johndoe@example.com"
                                    },
                                    "role_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utente creato con successo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Utente creato con successo"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dati non validi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Errore nella creazione dell'utente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Effettua il logout dell'utente",
                "description": "Permette di effettuare il logout, invalidando il token di accesso dell'utente e rimuovendo il token di ricordami.",
                "operationId": "userLogout",
                "responses": {
                    "200": {
                        "description": "Logout effettuato con successo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logout effettuato"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Utente non autenticato",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Utente non autenticato"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Role": {
                "title": "Role",
                "description": "Modello che rappresenta un ruolo utente",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Admin"
                    },
                    "users": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/User"
                        }
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "User",
                "description": "Modello che rappresenta un utente",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "cf": {
                        "type": "string",
                        "example": "RSSMRA85M01H501Z"
                    },
                    "name": {
                        "type": "string",
                        "example": "Mario"
                    },
                    "surname": {
                        "type": "string",
                        "example": "Rossi"
                    },
                    "birth_date": {
                        "type": "string",
                        "format": "date",
                        "example": "1985-01-01"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "mario.rossi@example.com"
                    },
                    "role_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "role": {
                        "$ref": "#/components/schemas/Role"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "JWT": {
                "type": "apiKey",
                "in": "header",
                "name": "Authorization",
                "description": "Token JWT per l'autenticazione"
            }
        }
    },
...
    "security": [
        {
            "JWT": []
        }
    ]
}